+++ src/Forms/test/Xamarin.Forms.Core.UnitTests/EasingTests.cs
@@ -7,15 +7,15 @@ namespace Xamarin.Forms.Core.UnitTests
 	public class EasingTests : BaseTestFixture
 	{
 		[Test]
-		public void Linear([Range(0, 10)] double input)
+		public void Linear([Range(0, 10)] float input)
 		{
 			Assert.AreEqual(input, Easing.Linear.Ease(input));
 		}
 
 		[Test]
-		public void AllRunFromZeroToOne([Values(0.0, 1.0)] double val)
+		public void AllRunFromZeroToOne([Values(0, 1.0f)] float val)
 		{
-			const double epsilon = 0.001;
+			const float epsilon = 0.001f;
 			Assert.True(Math.Abs(val - Easing.Linear.Ease(val)) < epsilon);
 			Assert.True(Math.Abs(val - Easing.BounceIn.Ease(val)) < epsilon);
 			Assert.True(Math.Abs(val - Easing.BounceOut.Ease(val)) < epsilon);
